Revert the changed logic of on_change_with_delivery_date introduced in changeset c685fe3b8b8a37d432fbdf68339456dafe66e519.
This change breaks the former behavior and causes a traceback in the workflow on step processing,
because it doesn't provide any shipping date when there is already a move without a delivery_date
(which can be very well the case when the sale line was created from code (i.e. purchase requests)
and not created from the client with other on_changes being triggered).

Traceback:
TypeError: unsupported operand type(s) for -: 'NoneType' and 'datetime.date'
on processing of a purchase line.